+++ config/cfg.go
@@ -70,6 +70,7 @@ func DefaultConfig() *Config {
 			RUMOriginIPHeader: "X-Forwarded-For",
 			Listen:            "localhost:9529",
 			RUMAppIDWhiteList: []string{},
+			PublicAPIs:        []string{},
 		},
 
 		DCAConfig: &dkhttp.DCAConfig{
@@ -579,6 +580,10 @@ func (c *Config) LoadEnvs() error {
 		c.Namespace = v
 	}
 
+	if v := datakit.GetEnv("ENV_ENABLE_ELECTION"); v != "" {
+		c.EnableElection = true
+	}
+
 	if v := datakit.GetEnv("ENV_GLOBAL_TAGS"); v != "" {
 		c.GlobalTags = ParseGlobalTags(v)
 	}
@@ -651,10 +656,6 @@ func (c *Config) LoadEnvs() error {
 		c.DefaultEnabledInputs = strings.Split(v, ",")
 	} else if v := datakit.GetEnv("ENV_ENABLE_INPUTS"); v != "" { // deprecated
 		c.DefaultEnabledInputs = strings.Split(v, ",")
-	}
-
-	if v := datakit.GetEnv("ENV_ENABLE_ELECTION"); v != "" {
-		c.EnableElection = true
 	}
 
 	if v := datakit.GetEnv("ENV_GIT_URL"); v != "" {