+++ io/makepoint_test.go
@@ -118,7 +118,7 @@ func TestNewPoint(t *testing.T) {
 		},
 
 		{
-			tname: "with disabled tag key",
+			tname: "with disabled tag key source",
 			name:  "abc",
 			mtype: datakit.Logging,
 			t:     map[string]string{"source": "s1"},
@@ -127,14 +127,6 @@ func TestNewPoint(t *testing.T) {
 		},
 
 		{
-			tname: "with disabled tag key",
-			name:  "abc",
-			mtype: datakit.Logging,
-			t:     map[string]string{"source": "s1"},
-			f:     map[string]interface{}{"f1": 123},
-			fail:  true,
-		},
-		{
 			tname: "with disabled field key",
 			name:  "abc",
 			mtype: datakit.Object,
@@ -142,6 +134,25 @@ func TestNewPoint(t *testing.T) {
 			f:     map[string]interface{}{"class": 123},
 			fail:  true,
 		},
+
+		{
+			tname: "with disabled tag key log_type",
+			name:  "abc",
+			mtype: datakit.Logging,
+			t:     map[string]string{"log_type": "s1"},
+			f:     map[string]interface{}{"f1": 123},
+			fail:  true,
+		},
+
+		{
+			tname: "with disabled field key log_type",
+			name:  "abc",
+			mtype: datakit.Logging,
+			t:     map[string]string{},
+			f:     map[string]interface{}{"log_type": 123},
+			fail:  true,
+		},
+
 		{
 			tname: "normal",