+++ plugins/inputs/container/k8s_pod.go
@@ -2,30 +2,16 @@ package container
 
 import (
 	"context"
-
-	// nolint:gosec
-	"crypto/md5"
-	"encoding/hex"
 	"fmt"
-	"strconv"
-	"strings"
 	"time"
 
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit"
-	"gitlab.jiagouyun.com/cloudcare-tools/datakit/config"
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit/io"
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit/plugins/inputs"
 	v1 "k8s.io/api/core/v1"
 )
 
 const k8sPodObjectName = "kubelet_pod"
-
-// const k8sPodName = "kubernetes_pod"
-
-const (
-	annotationPromExport  = "datakit/prom.instances"
-	annotationPromIPIndex = "datakit/prom.instances.ip_index"
-)
 
 func gatherPod(client k8sClientX, extraTags map[string]string) (k8sResourceStats, error) {
 	list, err := client.getPods().List(context.Background(), metaV1ListOption)
@@ -105,23 +91,11 @@ func exportPod(items []v1.Pod, extraTags tagsType) k8sResourceStats {
 		obj.time = time.Now()
 		res.set(defaultNamespace(item.Namespace), obj)
 
-		// discovery prom input
-		if config, ok := item.Annotations[annotationPromExport]; ok {
-			l.Info("k8s export, find prom export")
-			if !shouldForkInput(item.Spec.NodeName) {
-				l.Debugf("should not fork input, pod-nodeName:%s", item.Spec.NodeName)
-			} else {
-				d := &discovery{
-					name:      "prom",
-					cfg:       complatePromConfig(config, &items[idx]),
-					extraTags: extraTags,
-				}
-				if err := d.tryRunInput(); err != nil {
-					l.Warn(err)
-				}
-			}
+		if err := tryRunInput(&items[idx]); err != nil {
+			l.Warnf("failed to run input(discovery), %w", err)
 		}
 	}
+
 	return res
 }
 
@@ -218,7 +192,7 @@ func (*pod) Info() *inputs.MeasurementInfo {
 		},
 		Fields: map[string]interface{}{
 			"age":         &inputs.FieldInfo{DataType: inputs.Int, Unit: inputs.DurationSecond, Desc: "age (seconds)"},
-			"create_time": &inputs.FieldInfo{DataType: inputs.Int, Unit: inputs.UnknownUnit, Desc: "CreationTimestamp is a timestamp representing the server time when this object was created.(second)"},
+			"create_time": &inputs.FieldInfo{DataType: inputs.Int, Unit: inputs.TimestampSec, Desc: "CreationTimestamp is a timestamp representing the server time when this object was created.(second)"},
 			"restart":     &inputs.FieldInfo{DataType: inputs.Int, Unit: inputs.NCount, Desc: "The number of times the container has been restarted. (Depercated, use restarts)"},
 			"restarts":    &inputs.FieldInfo{DataType: inputs.Int, Unit: inputs.NCount, Desc: "The number of times the container has been restarted."},
 			"ready":       &inputs.FieldInfo{DataType: inputs.String, Unit: inputs.UnknownUnit, Desc: "container ready"},
@@ -227,106 +201,6 @@ func (*pod) Info() *inputs.MeasurementInfo {
 			"message":     &inputs.FieldInfo{DataType: inputs.String, Unit: inputs.UnknownUnit, Desc: "object details"},
 		},
 	}
-}
-
-// map[md5sum(cfg)] = nil.
-var discoveryInputsMap = make(map[string]interface{})
-
-type discovery struct {
-	name      string
-	cfg       string
-	extraTags map[string]string
-}
-
-func (d *discovery) tryRunInput() error {
-	creator, ok := inputs.Inputs[d.name]
-	if !ok {
-		return fmt.Errorf("invalid inputName")
-	}
-
-	md5str := md5sum(d.cfg)
-	if _, ok := discoveryInputsMap[md5str]; ok {
-		return nil
-	}
-
-	inputList, err := config.LoadInputConfig(d.cfg, creator)
-	if err != nil {
-		return err
-	}
-	discoveryInputsMap[md5str] = nil
-
-	l.Infof("discovery: add %s inputs, len %d", d.name, len(inputList))
-
-	// input run() 不受全局 election 影响
-	// election 模块运行在此之前，且其列表是固定的
-	g := datakit.G("kubernetes-autodiscovery")
-	for _, ii := range inputList {
-		if ii == nil {
-			l.Debugf("skip non-datakit-input %s", d.name)
-			continue
-		}
-
-		// switch ii.(type) {
-		// case inputs.OptionalInput:
-		// 	ii.SetTags(d.extraTags)
-		// }
-
-		func(name string, ii inputs.Input) {
-			g.Go(func(ctx context.Context) error {
-				l.Infof("discovery: starting input %s ...", name)
-				// main
-				ii.Run()
-				l.Infof("discovery: input %s exited", d.name)
-				return nil
-			})
-		}(d.name, ii)
-	}
-
-	return nil
-}
-
-func complatePromConfig(config string, podObj *v1.Pod) string {
-	podIP := podObj.Status.PodIP
-
-	func() {
-		indexStr, ok := podObj.Annotations[annotationPromIPIndex]
-		if !ok {
-			return
-		}
-		idx, err := strconv.Atoi(indexStr)
-		if err != nil {
-			l.Warnf("annotation prom.ip_index parse err: %s", err)
-			return
-		}
-		if !(0 <= idx && idx < len(podObj.Status.PodIPs)) {
-			l.Warnf("annotation prom.ip_index %d outrange, len(PodIPs) %d", idx, len(podObj.Status.PodIPs))
-			return
-		}
-		podIP = podObj.Status.PodIPs[idx].IP
-	}()
-
-	config = strings.ReplaceAll(config, "$IP", podIP)
-	config = strings.ReplaceAll(config, "$NAMESPACE", podObj.Namespace)
-	config = strings.ReplaceAll(config, "$PODNAME", podObj.Name)
-
-	return config
-}
-
-func shouldForkInput(nodeName string) bool {
-	if !datakit.Docker {
-		return true
-	}
-	// ENV NODE_NAME 在 daemonset.yaml 配置，是当前程序所在的 Node 名称
-	// Node 名称匹配，表示运行在同一个 Node，此时才需要 fork
-
-	// Node 名称为空属于 unreachable
-	return datakit.GetEnv("NODE_NAME") == nodeName
-}
-
-func md5sum(str string) string {
-	h := md5.New() //nolint:gosec
-	h.Write([]byte(str))
-	return hex.EncodeToString(h.Sum(nil))
 }
 
 //nolint:gochecknoinits