+++ plugins/inputs/jaeger/input.go
@@ -6,6 +6,7 @@ import (
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit"
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit/http"
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit/internal/trace"
+	iod "gitlab.jiagouyun.com/cloudcare-tools/datakit/io"
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit/plugins/inputs"
 )
 
@@ -59,6 +60,7 @@ func (*Input) SampleMeasurement() []inputs.Measurement {
 func (t *Input) Run() {
 	log = logger.SLogger(inputName)
 	log.Infof("%s input started...", inputName)
+	iod.FeedEventLog(&iod.Reporter{Message: "jaeger start ok, ready for collecting metrics.", Logtype: "event"})
 
 	if t.Tags != nil {
 		jaegerTags = t.Tags